SNES DRONE ROM V0.2
Had to do some changes regarding placement of routines in rom. Now, everything is in ROM bank 0, and tightly packed.
One more thing I could do is write to adress 00:4FEA and 00:4FEB. This is where the MCU firmware fetches the NMI vector address. It's an address in the free space of address bus A, so it might not be possible to write there. In that case, it just needs to be done when the ROM image is ready.